+++ RtCs.MathUtils/Geometry/Plane.cs
@@ -20,14 +20,14 @@
                 return Point;
             }
 
-            if (!Normal.Length2.AlmostEquals(1.0)) {
+            if (!Normal.Length2.AlmostEquals(1.0f)) {
                 Normal = Normal.Normalized;
             }
 
             return inPoint - (Vector3.Dot(inPoint - Point, Normal) * Normal);
         }
 
-        public double DistanceTo(Vector3 inPoint)
+        public float DistanceTo(Vector3 inPoint)
             => (Project(inPoint) - inPoint).Length;
     }
 }
